# [MRG] Expose samples belonging to a subcluster of Birch hierarchical tree