--- a/sklearn/cluster/birch.py
+++ b/sklearn/cluster/birch.py
@@ -250,6 +250,8 @@ class _CFSubcluster(object):
     linear_sum : ndarray, shape (n_features,), optional
         Sample. This is kept optional to allow initialization of empty
         subclusters.
+    samples_id : list, optional
+        Row number of samples belonging to the subcluster.
 
     Attributes
     ----------
@@ -274,21 +276,27 @@ class _CFSubcluster(object):
     sq_norm_ : ndarray
         Squared norm of the subcluster. Used to prevent recomputing when
         pairwise minimum distances are computed.
+
+    samples_id_ : list
+        Row number of samples belonging to the subcluster.
     """
-    def __init__(self, linear_sum=None):
+    def __init__(self, linear_sum=None, samples_id=None):
         if linear_sum is None:
             self.n_samples_ = 0
             self.squared_sum_ = 0.0
             self.linear_sum_ = 0
+            self.samples_id_ = []
         else:
             self.n_samples_ = 1
             self.centroid_ = self.linear_sum_ = linear_sum
             self.squared_sum_ = self.sq_norm_ = np.dot(
                 self.linear_sum_, self.linear_sum_)
+            self.samples_id_ = samples_id
         self.child_ = None
 
     def update(self, subcluster):
         self.n_samples_ += subcluster.n_samples_
+        self.samples_id_ += subcluster.samples_id_
         self.linear_sum_ += subcluster.linear_sum_
         self.squared_sum_ += subcluster.squared_sum_
         self.centroid_ = self.linear_sum_ / self.n_samples_
@@ -306,9 +314,10 @@ class _CFSubcluster(object):
         dot_product = (-2 * new_n) * new_norm
         sq_radius = (new_ss + dot_product) / new_n + new_norm
         if sq_radius <= threshold ** 2:
-            (self.n_samples_, self.linear_sum_, self.squared_sum_,
-             self.centroid_, self.sq_norm_) = \
-                new_n, new_ls, new_ss, new_centroid, new_norm
+            new_samples_id = self.samples_id_ + nominee_cluster.samples_id_
+            (self.n_samples_, self.samples_id_, self.linear_sum_,
+             self.squared_sum_, self.centroid_, self.sq_norm_) = \
+                new_n, new_samples_id, new_ls, new_ss, new_centroid, new_norm
             return True
         return False
 
@@ -475,8 +484,9 @@ class Birch(BaseEstimator, TransformerMixin, ClusterMixin):
         else:
             iter_func = _iterate_sparse_X
 
-        for sample in iter_func(X):
-            subcluster = _CFSubcluster(linear_sum=sample)
+        for row_id, sample in enumerate(iter_func(X)):
+            subcluster = _CFSubcluster(linear_sum=sample,
+                                       samples_id=[row_id])
             split = self.root_.insert_cf_subcluster(subcluster)
 
             if split: