Birch clustering constructs a hierarchical tree, using `_CFNode`, `_CFSubclusters` classes which expose attributes such as squared sum of samples, number of samples, centroid of samples etc. However, none of these allow to determine which samples are actually included in a given subcluster, which prevents exploring or doing additional computation on the constructed tree. 

This PR adds an additional attribute `sample_id_` to `_CFSubclusters` containing a list of row indices of the original `X` data array that belong to the current sub cluster.

#### Benchmarks 

This PR has a small performance and memory overhead  for the construction of the hierarchical tree (obtained after `Birch(n_clusters=None, compute_labels=False).fit(X)`. Since this phase is also faster than in the case when `compute_labels=True`  and  faster than the global clustering step that occurs when `n_clusters=<int>`, the overall performance impact is negligible. 

Here is an example of a benchmark,
```py
import numpy as np
from sklearn.cluster import Birch  # imported as Birch_PR for the code in this PR

np.random.seed(42)

X = np.random.randn(10000, 100)
```
before this PR we get,
```py
%timeit -n 5 Birch(n_clusters=None, compute_labels=False).fit(X)
%timeit -n 5 Birch(n_clusters=None, compute_labels=True).fit(X)
%timeit -n 5 Birch(n_clusters=100, compute_labels=True).fit(X)

1.07 s ± 5.93 ms per loop (mean ± std. dev. of 7 runs, 5 loops each)
1.74 s ± 3.09 ms per loop (mean ± std. dev. of 7 runs, 5 loops each)
8.96 s ± 16.8 ms per loop (mean ± std. dev. of 7 runs, 5 loops each)
```
after this PR we get,
```py
%timeit -n 5 Birch_PR(n_clusters=None, compute_labels=False).fit(X)
%timeit -n 5 Birch_PR(n_clusters=None, compute_labels=True).fit(X)
%timeit -n 5 Birch_PR(n_clusters=100, compute_labels=True).fit(X)
1.09 s ± 1.02 ms per loop (mean ± std. dev. of 7 runs, 5 loops each)
1.76 s ± 2.69 ms per loop (mean ± std. dev. of 7 runs, 5 loops each)
8.97 s ± 5.21 ms per loop (mean ± std. dev. of 7 runs, 5 loops each)
```
with mostly mostly the same run times.

cc. @MechCoder