@jnothman 2017-04-29
I've not looked yet, but as an out of core algorithm, i don't see how this
could use negligible memory in general.

On 29 Apr 2017 8:05 am, "Roman Yurchak" <notifications@github.com> wrote:

> Birch clustering constructs a hierarchical tree, using _CFNode,
> _CFSubclusters classes which expose attributes such as squared sum of
> samples, number of samples, centroid of samples etc. However, none of these
> allow to determine which samples are actually included in a given
> subcluster, which prevents exploring or doing additional computation on the
> constructed tree.
>
> This PR adds an additional attribute sample_id_ to _CFSubclusters
> containing a list of row indices of the original X data array that belong
> to the current sub cluster.
> Benchmarks
>
> This PR has a small performance and memory overhead for the construction
> of the hierarchical tree (obtained after Birch(n_clusters=None,
> compute_labels=False).fit(X). Since this phase is also faster than in the
> case when compute_labels=True and faster than the global clustering step
> that occurs when n_clusters=<int>, the overall performance impact is
> negligible.
>
> Here is an example of a benchmark,
>
> import numpy as npfrom sklearn.cluster import Birchfrom freediscovery.externals.birch import Birch as Birch_PR
>
> np.random.seed(42)
>
> X = np.random.randn(10000, 100)
>
> before this PR we get,
>
> %timeit -n 5 Birch(n_clusters=None, compute_labels=False).fit(X)%timeit -n 5 Birch(n_clusters=None, compute_labels=True).fit(X)%timeit -n 5 Birch(n_clusters=100, compute_labels=True).fit(X)
> 1.07 s ± 5.93 ms per loop (mean ± std. dev. of 7 runs, 5 loops each)1.74 s ± 3.09 ms per loop (mean ± std. dev. of 7 runs, 5 loops each)8.96 s ± 16.8 ms per loop (mean ± std. dev. of 7 runs, 5 loops each)
>
> after this PR we get,
>
> %timeit -n 5 Birch_PR(n_clusters=None, compute_labels=False).fit(X)%timeit -n 5 Birch_PR(n_clusters=None, compute_labels=True).fit(X)%timeit -n 5 Birch_PR(n_clusters=100, compute_labels=True).fit(X)1.09 s ± 1.02 ms per loop (mean ± std. dev. of 7 runs, 5 loops each)1.76 s ± 2.69 ms per loop (mean ± std. dev. of 7 runs, 5 loops each)8.97 s ± 5.21 ms per loop (mean ± std. dev. of 7 runs, 5 loops each)
>
> with mostly mostly the same run times.
>
> cc. @MechCoder <https://github.com/MechCoder>
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/scikit-learn/scikit-learn/pull/8808
> Commit Summary
>
>    - Add samples_id_ attribute to Birch _CFSubcluster
>
> File Changes
>
>    - *M* sklearn/cluster/birch.py
>    <https://github.com/scikit-learn/scikit-learn/pull/8808/files#diff-0>
>    (22)
>    - *M* sklearn/cluster/tests/test_birch.py
>    <https://github.com/scikit-learn/scikit-learn/pull/8808/files#diff-1>
>    (43)
>
> Patch Links:
>
>    - https://github.com/scikit-learn/scikit-learn/pull/8808.patch
>    - https://github.com/scikit-learn/scikit-learn/pull/8808.diff
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/scikit-learn/scikit-learn/pull/8808>, or mute the
> thread
> <https://github.com/notifications/unsubscribe-auth/AAEz64CQFsEsMCNCW7oO8L7lSB9K6FJLks5r0mLFgaJpZM4NMBPm>
> .
>