@MechCoder 2017-05-04
Sorry for the delayed reply, why do you need this? If I recall correctly, one of the selling points was to do instance reduction without keeping track of which samples belong to which subcluster.

You can get an approximate answer by just doing minimum distances with the subcluster centers?