@MechCoder 2017-05-04
I'm okay with making it optional provided you have a particular use case for it.