@jnothman 2017-05-04
I'm inclined to agree that it seems against the spirit of the technique

On 5 May 2017 2:38 am, "Manoj Kumar" <notifications@github.com> wrote:

> I'm okay with making it optional provided you have a particular use case
> for it.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/scikit-learn/scikit-learn/pull/8808#issuecomment-299240941>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAEz6wUjLldohB5m0TlPZulk7OffXk6Vks5r2f8dgaJpZM4NMBPm>
> .
>